OpenAPI resolver as a package.
The openapi resolver is now a reusable package with
an embedded script. Added tests and travis.yml
