+++ .travis.yml
@@ -1,0 +1,9 @@
+# ref: https://docs.travis-ci.com/user/languages/python
+language: python
+python:
+  - "3.6"
+  #- "nightly" # points to the latest development branch e.g. 3.6-dev
+# command to install dependencies
+install: "pip install -r requirements.txt"
+# command to run tests
+script: nosetests scripts/openapi_resolver.py --with-coverage --cover-package=openapi_resolver
