+++ requirements.txt
@@ -2,3 +2,6 @@ connexion == 1.1.15
 python_dateutil == 2.6.0
 setuptools >= 21.0.0
 pyOpenSSL
+coverage>=4.0.3
+nose>=1.3.7
+
